requirements.txt
tensorflow==2.17.0
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
opencv-python
fastapi
uvicorn
onnx
onnxruntime
tf2onnx
scikit-learn
numpy
matplotlib
pandas
Pillow
keras==3.4.1